v0.0.2: Pinning Summit + initial feedback
This version is  following spec drafted during Pinning Summit
where we switched API to operate on "pin objects"
with some improvements based on initial feedback and discussions
received from the community this week.

This is by no means the finished spec: goal is to close some of topics
and kick-off discussions on remaining items.