docs: case-sensitive Pin.name filter
This removes ambiguity, as noted in
https://github.com/ipfs/pinning-services-api-spec/issues/45#issuecomment-682608214

The rationale for going with case-sensitive is that it is more
performance out of the box, does not require any additional indexes or
optimizations.

Case-insensitive comes with a risk for bugs when a search results in table scan
even when there is an index on the name column.