feat: fast name search by default
This changes the default behavior of the `name` filter from
"case-insensitive, partial match" to "case-sensitive, exact (full)
match" and adds `match` parameter that enables opt-in into `fuzzy`
strategy only when performance is not an object.

The rationale for this change that fuzzy search is a bad default, does
not scale well and in most cases users and developers expect exact
match. We are moving away from slow defaults in go-ipfs, and since
`ipfs pin local` API will be inspired by this spec, we don't want to
re-introduce bad choices into IPFS ecosystem.

Developers who need fuzzy search, can always opt-in into slower, but
more flexible text matching strategy by passing `match` parameter, but
the default should be the fastest option available,
and in  this case it is the exact match.